+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Alter button styling on `LitDatePicker`.
 - Alter `LitCreditsScreen`'s back button color.
 - Add `resizeToAvoidBottomInset` member on `LitTabView` to enforce overlapping according to [Material Design's guidelines](https://material.io/components/app-bars-bottom#behavior).
+- Implement `DialogActionButton` on `DiscardDraftDialog`.
 
 ## 1.0.4